Add files via upload
After the CNNs are trained we test their power with the collaboration of the Classify_Areas module and the images on SampleK folder that you can download from my google drive, the link is on Readme.md. Also before running the modes that will be combined for the predictions must be specified(1-20). You are welcome to try any combination you wish to test , but make sure the Fully Connected NN is trained with this combination. The FC-NN is trained by the FC_Classifier_With_Probabilties.py module