Add files via upload
In every Model that we train we save the history of the last 100 Epochs in a csv file in the same folder that the model is saved. History contains Accuracy , Validation Accuracy , 
Loss and Validation Loss. With this module you can see the mean value , variance and the peaks for those variables . This is for model testing purposes.